Fix function String_getToken testing EOF in string: EOF constant isn't a real character and therefore can't exist in strings; escalate type of 'unsigned short int' typed variables into 'unsigned int' in function String_getToken

+++ StringUtils.h
@@ -65,7 +65,7 @@ char** String_split(const char* s, char sep, int* n);
 
 void String_freeArray(char** s);
 
-char* String_getToken(const char* line, const unsigned short int numMatch);
+char* String_getToken(const char* line, const unsigned int numMatch);
 
 char* String_readLine(FILE *f);
